+++ app/src/main/java/me/androidbox/flicks/moviedetail/MovieDetailModelImp.java
@@ -16,8 +16,7 @@ import timber.log.Timber;
  */
 
 public class MovieDetailModelImp implements MovieDetailModelContract {
-    @Inject
-    FlicksMovieService mFlicksMovieService;
+    @Inject FlicksMovieService mFlicksMovieService;
 
     public MovieDetailModelImp() {
         DaggerInjector.getsAppComponent().inject(MovieDetailModelImp.this);
@@ -28,7 +27,7 @@ public class MovieDetailModelImp implements MovieDetailModelContract {
     }
 
     @Override
-    public void getMovieDetail(final Long movie_id, GetMovieDetailListener getMovieDetailListener) {
+    public void getMovieDetail(final int movie_id, GetMovieDetailListener getMovieDetailListener) {
         mFlicksMovieService.getMovieDetail(movie_id, Constants.API_KEY).enqueue(new Callback<MovieDetail>() {
             @Override
             public void onResponse(Call<MovieDetail> call, Response<MovieDetail> response) {
